NeuralOperators

Documentation for NeuralOperators.

NeuralOperators.FourierOperator

NeuralOperators.SpectralConv

Layers

Spectral convolutional layer

\[F(s) = \mathcal{F} \{ v(x) \} \\ F'(s) = g(F(s)) \\ v'(x) = \mathcal{F}^{-1} \{ F'(s) \}\]

where $v(x)$ and $v'(x)$ denotes input and output function, $\mathcal{F} \{ \cdot \}$, $\mathcal{F}^{-1} \{ \cdot \}$ are Fourier transform, inverse Fourier transform, respectively. Function $g$ is a linear transform for lowering Fouier modes.

NeuralOperators.SpectralConv — Type

SpectralConv(
    ch, modes, σ=identity;
    init=c_glorot_uniform, permuted=false, T=ComplexF32
)

Arguments

ch: Input and output channel size, e.g. 64=>64.

modes: The Fourier modes to be preserved.

σ: Activation function.

permuted: Whether the dim is permuted. If permuted=true, layer accepts data in the order of (ch, ..., batch), otherwise the order is (..., ch, batch).

Example

julia> SpectralConv(2=>5, (16, ))
SpectralConv(2 => 5, (16,), σ=identity, permuted=false)

julia> using Flux

julia> SpectralConv(2=>5, (16, ), relu)
SpectralConv(2 => 5, (16,), σ=relu, permuted=false)

julia> SpectralConv(2=>5, (16, ), relu, permuted=true)
SpectralConv(2 => 5, (16,), σ=relu, permuted=true)

Reference: Fourier Neural Operator for Parametric Partial Differential Equations

Fourier operator layer

\[v_{t+1}(x) = \sigma(W v_t(x) + \mathcal{K} \{ v_t(x) \} )\]

where $v_t(x)$ is the input function for $t$-th layer and $\mathcal{K} \{ \cdot \}$ denotes spectral convolutional layer. Activation function $\sigma$ can be arbitrary non-linear function.

NeuralOperators.FourierOperator — Type

FourierOperator(ch, modes, σ=identity; permuted=false)

Arguments

ch: Input and output channel size for spectral convolution, e.g. 64=>64.

modes: The Fourier modes to be preserved for spectral convolution.

σ: Activation function.

permuted: Whether the dim is permuted. If permuted=true, layer accepts data in the order of (ch, ..., batch), otherwise the order is (..., ch, batch).

Example

julia> FourierOperator(2=>5, (16, ))
FourierOperator(2 => 5, (16,), σ=identity, permuted=false)

julia> using Flux

julia> FourierOperator(2=>5, (16, ), relu)
FourierOperator(2 => 5, (16,), σ=relu, permuted=false)

julia> FourierOperator(2=>5, (16, ), relu, permuted=true)
FourierOperator(2 => 5, (16,), σ=relu, permuted=true)

Reference: Fourier Neural Operator for Parametric Partial Differential Equations