Index

NeuralOperators.FourierOperator

NeuralOperators.SpectralConv

NeuralOperators.FourierNeuralOperator

NeuralOperators.MarkovNeuralOperator

Layers

Spectral convolutional layer

\[F(s) = \mathcal{F} \{ v(x) \} \\ F'(s) = g(F(s)) \\ v'(x) = \mathcal{F}^{-1} \{ F'(s) \}\]

where $v(x)$ and $v'(x)$ denotes input and output function, $\mathcal{F} \{ \cdot \}$, $\mathcal{F}^{-1} \{ \cdot \}$ are Fourier transform, inverse Fourier transform, respectively. Function $g$ is a linear transform for lowering Fouier modes.

NeuralOperators.SpectralConv — Type

SpectralConv(
    ch, modes;
    init=c_glorot_uniform, permuted=false, T=ComplexF32
)

Arguments

ch: Input and output channel size, e.g. 64=>64.

modes: The Fourier modes to be preserved.

permuted: Whether the dim is permuted. If permuted=true, layer accepts data in the order of (ch, ..., batch), otherwise the order is (..., ch, batch).

Example

julia> SpectralConv(2=>5, (16, ))
SpectralConv(2 => 5, (16,), permuted=false)

julia> SpectralConv(2=>5, (16, ), permuted=true)
SpectralConv(2 => 5, (16,), permuted=true)

Reference: Fourier Neural Operator for Parametric Partial Differential Equations

Fourier operator layer

\[v_{t+1}(x) = \sigma(W v_t(x) + \mathcal{K} \{ v_t(x) \} )\]

where $v_t(x)$ is the input function for $t$-th layer and $\mathcal{K} \{ \cdot \}$ denotes spectral convolutional layer. Activation function $\sigma$ can be arbitrary non-linear function.

NeuralOperators.FourierOperator — Type

FourierOperator(ch, modes, σ=identity; permuted=false)

Arguments

ch: Input and output channel size for spectral convolution, e.g. 64=>64.

modes: The Fourier modes to be preserved for spectral convolution.

σ: Activation function.

permuted: Whether the dim is permuted. If permuted=true, layer accepts data in the order of (ch, ..., batch), otherwise the order is (..., ch, batch).

Example

julia> FourierOperator(2=>5, (16, ))
FourierOperator(2 => 5, (16,), σ=identity, permuted=false)

julia> using Flux

julia> FourierOperator(2=>5, (16, ), relu)
FourierOperator(2 => 5, (16,), σ=relu, permuted=false)

julia> FourierOperator(2=>5, (16, ), relu, permuted=true)
FourierOperator(2 => 5, (16,), σ=relu, permuted=true)

Reference: Fourier Neural Operator for Parametric Partial Differential Equations

Models

Fourier neural operator

NeuralOperators.FourierNeuralOperator — Function

FourierNeuralOperator(;
    ch=(2, 64, 64, 64, 64, 64, 128, 1),
    modes=(16, ),
    σ=gelu
)

Fourier neural operator learns a neural operator with Dirichlet kernel to form a Fourier transformation. It performs Fourier transformation across infinite-dimensional function spaces and learns better than neural operator.

Reference: Fourier Neural Operator for Parametric Partial Differential Equations

Markov neural operator

NeuralOperators.MarkovNeuralOperator — Function

MarkovNeuralOperator(;
    ch=(1, 64, 64, 64, 64, 64, 1),
    modes=(24, 24),
    σ=gelu
)

Markov neural operator learns a neural operator with Fourier operators. With only one time step information of learning, it can predict the following few steps with low loss by linking the operators into a Markov chain.

Reference: Markov Neural Operators for Learning Chaotic Systems